+++ AdaptiClass/frontend/src/authentication/AuthenticationGuard.js
@@ -1,6 +1,6 @@
 import { withAuthenticationRequired } from "@auth0/auth0-react";
 import React from "react";
-import { AppLoader } from "../pages/AppLoader";
+import { AppLoader } from "../pages/helperScreens/AppLoader";
 
 export const AuthenticationGuard = ({ component }) => {
   const Component = withAuthenticationRequired(component, {
