+++ AdaptiClass/frontend/src/mockRequests/assignments.json
