--- AdaptiClass/frontend/src/pages/Assignments.js
@@ -1,1 +1,0 @@
-//I think this is nothin.
