+++ AdaptiClass/frontend/src/pages/Home.js
@@ -7,7 +7,7 @@ import timeData from '../time.json';
 import Upcoming from '../components/Upcoming';
 import toDoData from '../upcoming.json';
 import Engagement from '../components/Engagement'; 
-import engagementData from '../engagement.json';
+import engagementData from '../mockRequests/engagement.json';
 
 const Content = styled.div`
   display: flex;
