+++ AdaptiClass/frontend/src/pages/helperScreens/NewUser.js
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import styled from "styled-components";
-import { RocketIcon } from "../assets/Icons";
+import { RocketIcon } from "../../assets/Icons";
 import { useAuth0 } from "@auth0/auth0-react";
 import axios from "axios";
 import { useNavigate } from "react-router-dom";
